Just use `fast-deep-equal`.
I only ever use JSON, but if anyone ever does adopt Proof and want to
compare `Date` or `RegExp` I don't want to have to preach to them. It
would be enough for me to walk away from adoption if something I
expected to be able to do, like compare `undefined`, where to be
unsupported by a library.

See #28.